Add tanh mech and revert to local run only
This commit

1. renames a couple files, since the script vs. mech differences are
obvious

2. Adds `tanh` types of the synaptic mechanisms for speed comparison to
the native `COBAHH` type benchmarks, which rely on conditionals in the
synaptic mechanisms. Conditionals may offer a speed penalty when used in
mechanisms, and this way we can compare for that penalty at scale.

3. The scripts were reverted to much smaller simulations so you can run
them on your local computer, instead of the cluster. This is due to a
bug in the DynaSim `debug` branch right now, and is very easy to change.

 Changes to be committed:
	renamed:    benchmark_runscript_COBAHH_nodisk.m -> benchmark_COBAHH_nodisk.m
	renamed:    benchmark_runscript_COBAHH_yesdisk.m -> benchmark_COBAHH_yesdisk.m
	new file:   benchmark_COBAHHtanh_nodisk.m
	new file:   models/iAMPAtanhBM.mech
	new file:   models/iGABAatanhBM.mech